+++ src/main/java/com/Teknisi/model/TeknisiPhoto.java
@@ -24,7 +24,7 @@ public class TeknisiPhoto implements Serializable{
 	@NotNull(message = "Teknisi ID cannot be null")
 	@Max(value = 1000, message = "ID should not be greater than 1000")
 	@ApiModelProperty(notes = "teknisi_id of the Teknisi Photo", name = "teknisi_id", required = true, example = "1")
-	private int teknisi_id;
+	private Long teknisi_id;
 	
 	@ApiModelProperty(hidden = true)
 	private String file_type;
@@ -55,7 +55,7 @@ public class TeknisiPhoto implements Serializable{
 
 	public TeknisiPhoto(
 			@NotNull(message = "Teknisi Photo ID cannot be blank") @Max(value = 1000, message = "ID should not be greater than 1000") Long id,
-			@NotNull(message = "Teknisi ID cannot be null") @Max(value = 1000, message = "ID should not be greater than 1000") int teknisi_id,
+			@NotNull(message = "Teknisi ID cannot be null") @Max(value = 1000, message = "ID should not be greater than 1000") Long teknisi_id,
 			String file_type, String name, String images) {
 		super();
 		this.id = id;
@@ -73,11 +73,11 @@ public class TeknisiPhoto implements Serializable{
 		this.id = id;
 	}
 
-	public int getTeknisi_id() {
+	public Long getTeknisi_id() {
 		return teknisi_id;
 	}
 
-	public void setTeknisi_id(int teknisi_id) {
+	public void setTeknisi_id(Long teknisi_id) {
 		this.teknisi_id = teknisi_id;
 	}
